3. Тесты для определения расстояния между точками. [#343992]

+++ src/ru/job4j/Point.java
@@ -2,7 +2,7 @@ package ru.job4j;
 
 public class Point {
 
-    public static double distance(int x1, int y1, int x2, int y2) {
+    public static double distance(double x1, double y1, double x2, double y2) {
         double x3 = x2 - x1;
         double y3 = y2 - y1;
         double x3sq = Math.pow(x3, 2);
